Rename package pastishf
